Define quests and check for location enter
Added a simple definition for quests and check the stream of location
updates against any of the targets.  Currently, a quest is activated
only when you enter a quest zone initially.  Jumping in and out of the
circle won't retrigger until you enter a different quest zone.

Quests can additionally specify color and stroke width, although there
are sane defaults if not present.