Use AJAX for file loading on device
window.cordova appears to be a good indicator that a phonegap app
is running on a device, not in the browser. Device apps are able
to request files in the app root via ajax, similarly to browser.
The FileReader api seems to only have access to the "Documents"
directory.

I chose to keep the file reader abstraction even though all
platforms will use the same mechanism as it might be useful
to generate all requests from a single module.