Hide zone locations. Add zone location hint.
* Zones are now not shown on the map, but users are given a hint as to
their location. The marker for the hint starts as a circle with
no radius, then grows over 1 second to the full size of the zone.
The color of the marker also becomes more and more yellow. When the
animation completes, the marker disappears again, and the animation
is triggered 30 seconds later. The timing between each hint cycle,
as well as the duration of the hint, can be configured.
* Remove the stroke for the unstarted marker to better match the hint
style.