Show extra images on quest completion.
When a user finishes a task we would like to show extra images. These images are
geotagged and will show up on the map as clickable/tapable markers. Upon
clicking a marker, the image will appear in a popup similarly to the way quests
are displayed though you will not have to be in the geographic bounds to see
this image. A new set of functions are appended to the quest finsihed stream to
see which zone has been completed. Extra data, which is part of the quest.json
data is them pulled off of this and used to create markers associated with the
proper images.