Enable mock clicking inside zones
* When the feature to show quests for completed zones via clicking
of a zone marker was added, click events on zones were captured and did
not propagate to the map. This prevented the user from moving into a
zone with the mock location setting enabled. To fix this, the zone
click event isn't assigned until the zone is finished, and it's removed
if the zone's status changes back to unstarted or in progress.

Closes #28