Add new Fairmount Water Works quest and zones.
* Add new content from Victoria. This process involved converting
tif files to jpegs, reworking the directory structure to be a bit
simpler and to avoid folders with spaces in the names, and adding
a json file to each zone that contains configuration and image
metadata for that zone. Image names were left as-is. There are some
images that weren't used in the storyboards document, but I checked
them in anyway in case we need to add them later.
* Ignore *.tif. This allows us to keep the original files (*.tifs)
next to the jpeg files that we are using in case we need to generate
new images from the originals.
* Tell git not to try and parse image files to create a diff. This
ruins the images.

Connects to #35