Add zones and zone status to quests page.
The quest page needs to show a list of quests and finished zones in those
quests. For right now, since we only have one quest, we just show all the zones.
Finished zones can be accessed by clicking (tapping) the row and are designated
with a checkmark. Unfinished zones are shown only as a text name and are not
interactive.

 * `map.init()` returns the questManager
 * questManager passed to ui and pageManager to access zone data.
 * Added pageManager (`pages.js`) which handles loading data into menu pages.