Add swipe event to cards.
We want users to be able to navigate through the card decks using swipe
gestures. This can be accomplished using jQuery Mobile. We add listener to the
card holder for left and right swipes and use those to trigger click events on
the links in the header or footer of the card. This minimizes the logic on the
swipe handler since the links are already properly wired up to the proper
transition logic. One issue that was encounted was the fact that jquery mobile
swipe events seem to conflict with ```-webkit-overflow-scrolling: touch``` so
the `scrollable` class needed to be removed from the card captions. This should
not be an issue since the card content is in most cases only a few lines and
since we will be adding the ability to hide it shortly.