Adjust image CSS to support varying aspect ratios.
The zone images are of varying aspect ratios and, in most cases,
the full image was not being displayed in the app. To improve
how the images are displayed, adjust the card style to scale the
image to fit within the card. Also, change the background color
of the card so that any space not filled by the image is black.

Connects to #49.