Support use of video.
Instead of image tags, in some places we want to show videos. We'll use the
HTML5 video tag element and the name of the video will be the full filename so
that if we get other formats we will be able to use them without other logic. We
will mandate the addition of a poster frame to act as a placeholder.