Deal with iOS highjacking events on video.
It appears that HTML5 video playing on iOS causes problems with touch
events. The video seems to take over and surpress all other events, such
as the swipe event that we were using for navigation. I got around this
by creating a poster div that is swipable. Tapping this div shows the
video. When the video is compeleted or paused the poster returns to
allow further navigation.