Add directional arrow/compass.
We want to give the user feedback about the direction of movement since
the map does not reorient. We do so by adding a small circle with
navigation arrow in it that turns according to the devices deviation
from magnetic north. This is done by applying a rotation transformation
to an SVG.