Disable cloud based-storage.
We're not storing anything on the app so disable cloud based-storage
which is on by default. In iOS this tries to store things to iCloud and
the builds have been throwing warnings about this. Since this may end up
in the app store we want to reduce the surface area for problems in the
approval process.