Fix fontello font config.
Fontello allows you to download just the icon fonts that you need and
provides a config.json file to help with this. That config file can be
used to restore the picked fonts in the fontello ui and then add icons
to it. Without this, you have to manually pick each font you want again.
I found that we were only using two fontello fonts and added them to the
project along with a third I want for the navigation arrow. Removed the
old font CSS and fonts so we don't load things we don't need.