Create 'font' directory.
Fontello, by default looks for fonts in a directory called 'font' rather
than 'fonts' which is what the bootstrap css looks for. We've been
manually altering the css that comes from fontello to allow us to
centralize our fonts. This wastes time when updating or adding fonts
because we can't just use the css that fontello gives us and inevitably
we waste time debugging the fonts that fail to load if we forget to
manually change this. Since there's no real harm or overhead required to
have a separte font directory for these libraries we do this to make
future changes to the font stack easier.