Enable global instruction card & remove captions.
We want a global instruction card to show the first and only the first
time a user enters a zone. We produce a global flag to handle the
initial showing of this and append it as a card to the zone that is
activated. After this first time, the instruction card is never shown
again. We also remove the captions and replace them with navigation
instructions which can no longer be hidden. We also remove the ability
to swipe the zone away by exiting right through the first card.