Enable alignment message & rework css.
The first slide should show and alignment message. To deal with this and
possible contradictions with the existing audio player header, we
activate the alignment message as a div that sits overtop of the
standard audio mechanism and controls. Swiping to the left on the first
card removes this div and reveals the normal controls. Reworked and
cleaned up some CSS in the process.