Play audio when changing to the second slide.
Since the first card has instructions and the user can't control the
audio from it, we start the audio playing after swiping to the second
card. A strange issue came about testing on Chrome. The audio would
queue but be blocked by a previous attempt to play the file. This
appears to only properly cancel by using the `window.stop()` method. For
good measure a cache busting token is affixed to the audio source's
`src` attribute to ensure we don't have issues with stale data. This
shouldn't be a problem as all the files will be hosted locally anyway.