Adjust audio playing logic.
If the user has paused the audio or if the audio has reached the end we
want to appropriately handle the messages and playing when moving
between cards. Since we allow the user to revisit the first card (that
triggers the audio playing) we want to ensure we respect the paused
state assuming they have already moved to the next card and manually
stopped the player. We also want to ensure we properly update the button
text. To handle this better the button text is wired not to the events
on the button but to the actual audio source. So if the audio is paused
or the audio ends, the text changes appropriately.