Don't watch tiles or quests directory.
The tiles directory has enough files that it can cause trouble with
` grunt watch`. We rarely change the tiles anyway and certainly not
during active coding where we need a fast feedback loop so we can
exclude these. Quests contain the basic data for the application and are
also rarely changed. Quests are also generated for the application by a
script and so aren't a good case for normal file watching.