Resize quote text on device rotation.
When the device changes orientation, the text may no longer fit. Since
rerendering takes some time after the rotation, we add a setTimeout and wait a
second before resetting the font size. This also adds a smaller font increment
and accounts for the size of the instruction bar at the bottom of the screen.