Apply targeted scaling to intro text.
We want to ensure that the intro text does not get too large and that
users can move on to the next screen even on smaller devices. Extract
the text scaling function used previously on the historic context text
of cards and make it generically useful to any text and container.